LE: nonexclusive pointer update small code cleanup

+++ MyGUIEngine/src/MyGUI_PopupMenu.cpp
@@ -75,7 +75,7 @@ namespace MyGUI
 		StaticImagePtr image = _button->getStaticImage();
 		if ( null == image ) return;
 		if (image->getItemResource()) {
-			static const int CountIcons = 2;
+			static const size_t CountIcons = 2;
 			static const char * IconNames[CountIcons + 1] = {"None", "Popup", ""};
 			if (_index >= CountIcons) _index = CountIcons;
 			image->setItemName(IconNames[_index]);
